+++ CHANGELOG.md
@@ -12,6 +12,9 @@ From version 2.6.0, the sections in this file adhere to the [keep a changelog](h
 * [#2092](https://github.com/Shopify/shopify-cli/pull/2092): Fix `RootHelper` parse logic to support options with an equal (e.g.: `option=value`)
 * [#2089](https://github.com/Shopify/shopify-cli/pull/2089): Use javy version 0.2.0
 
+### Removed
+* [#2102](https://github.com/Shopify/shopify-cli/pull/2102): Remove AssemblyScript as a supported script language.
+
 ## Version 2.12.0
 ### Added
 * [#1866](https://github.com/Shopify/shopify-cli/pull/1866): Enforce git dependency